+++ packages/control-statement/__tests__/for.spec.tsx
@@ -25,6 +25,42 @@ const TestForObject = (props: { obj: Record<string, number> }) => {
   );
 };
 
+const TestForMap = (props: { map: Map<string, number> }) => {
+  return (
+    <For of={props.map}>
+      <Empty>{() => 'empty'}</Empty>
+      {(item, { index, key }) => <i key={index}>{item}</i>}
+    </For>
+  );
+};
+
+const TestForWeakMap = (props: { map: WeakMap<object, number> }) => {
+  return (
+    <For of={props.map}>
+      <Empty>{() => 'empty'}</Empty>
+      {(item, { index, key }) => <i key={index}>{item}</i>}
+    </For>
+  );
+};
+
+const TestForSet = (props: { set: Set<number> }) => {
+  return (
+    <For of={props.set}>
+      <Empty>{() => 'empty'}</Empty>
+      {(item, { index, key }) => <i key={index}>{item}</i>}
+    </For>
+  );
+};
+
+const TestForWeakSet = (props: { set: WeakSet<object> }) => {
+  return (
+    <For of={props.set}>
+      <Empty>{() => 'empty'}</Empty>
+      {(item, { index, key }) => <i key={index}>{item}</i>}
+    </For>
+  );
+};
+
 describe('for element', function() {
   const app = mount(<TestFor list={[1, 2, 3]} />);